reasons for fix
we don't need to convert _amount to some other unit since already in ProxyActions contract, the _amount is msg.value(which is wei :10^18) that's why even when to18ConversionFactor was 1, everything worked fine. Therefore I removed needless conversion factors and decimals which comes from aXDCcAddress. This ankrColAdapter doesn't care about aXDCc when it deals with _share.